dependencyinjection/base-models/src/main/java/com/ford/turbo/aposb/common/filter/TimingFilter.java
package com.ford.turbo.aposb.common.filter;

import java.io.IOException;

import javax.servlet.FilterChain;
import javax.servlet.ServletException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.stereotype.Component;
import org.springframework.web.filter.OncePerRequestFilter;

@Component
public class TimingFilter extends OncePerRequestFilter {

    private static final Logger LOGGER = LoggerFactory.getLogger(TimingFilter.class.getName());

    @Override
    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
        long startTime = System.currentTimeMillis();
        filterChain.doFilter(request, response);
        LOGGER.info("Total Execution time took {} ms", System.currentTimeMillis() - startTime);
    }
}
